+++ sdk/authorization/mgmt/src/main/java/com/azure/resourcemanager/authorization/implementation/RoleAssignmentsImpl.java
@@ -1,0 +1,103 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.resourcemanager.authorization.implementation;
+
+import com.azure.core.http.rest.PagedFlux;
+import com.azure.core.http.rest.PagedIterable;
+import com.azure.core.management.exception.ManagementException;
+import com.azure.resourcemanager.authorization.RoleAssignment;
+import com.azure.resourcemanager.authorization.RoleAssignments;
+import com.azure.resourcemanager.authorization.models.RoleAssignmentInner;
+import com.azure.resourcemanager.authorization.models.RoleAssignmentsInner;
+import com.azure.resourcemanager.resources.fluentcore.arm.collection.implementation.CreatableResourcesImpl;
+import com.azure.resourcemanager.resources.fluentcore.model.HasInner;
+import reactor.core.publisher.Mono;
+
+/** The implementation of RoleAssignments and its parent interfaces. */
+class RoleAssignmentsImpl extends CreatableResourcesImpl<RoleAssignment, RoleAssignmentImpl, RoleAssignmentInner>
+    implements RoleAssignments, HasInner<RoleAssignmentsInner> {
+    private final GraphRbacManager manager;
+
+    RoleAssignmentsImpl(final GraphRbacManager manager) {
+        this.manager = manager;
+    }
+
+    @Override
+    protected RoleAssignmentImpl wrapModel(RoleAssignmentInner roleAssignmentInner) {
+        if (roleAssignmentInner == null) {
+            return null;
+        }
+        return new RoleAssignmentImpl(roleAssignmentInner.name(), roleAssignmentInner, manager());
+    }
+
+    @Override
+    public RoleAssignmentImpl getById(String objectId) {
+        return (RoleAssignmentImpl) getByIdAsync(objectId).block();
+    }
+
+    @Override
+    public Mono<RoleAssignment> getByIdAsync(String id) {
+        return inner()
+            .getByIdAsync(id)
+            .onErrorResume(ManagementException.class, e -> Mono.empty())
+            .map(
+                roleAssignmentInner ->
+                    new RoleAssignmentImpl(roleAssignmentInner.name(), roleAssignmentInner, manager()));
+    }
+
+    @Override
+    public RoleAssignmentImpl getByScope(String scope, String name) {
+        return (RoleAssignmentImpl) getByScopeAsync(scope, name).block();
+    }
+
+    @Override
+    public PagedFlux<RoleAssignment> listByScopeAsync(String scope) {
+        return inner().listForScopeAsync(scope, null).mapPage(roleAssignmentInner -> wrapModel(roleAssignmentInner));
+    }
+
+    @Override
+    public PagedIterable<RoleAssignment> listByScope(String scope) {
+        return wrapList(inner().listForScope(scope, null));
+    }
+
+    @Override
+    public Mono<RoleAssignment> getByScopeAsync(String scope, String name) {
+        return inner()
+            .getAsync(scope, name)
+            .onErrorResume(ManagementException.class, e -> Mono.empty())
+            .map(
+                roleAssignmentInner ->
+                    new RoleAssignmentImpl(roleAssignmentInner.name(), roleAssignmentInner, manager()));
+    }
+
+    @Override
+    protected RoleAssignmentImpl wrapModel(String name) {
+        return new RoleAssignmentImpl(name, new RoleAssignmentInner(), manager());
+    }
+
+    @Override
+    public Mono<RoleAssignment> deleteByIdAsync(String id) {
+        return inner()
+            .deleteByIdAsync(id)
+            .onErrorResume(ManagementException.class, e -> Mono.empty())
+            .map(
+                roleAssignmentInner ->
+                    new RoleAssignmentImpl(roleAssignmentInner.name(), roleAssignmentInner, manager()));
+    }
+
+    @Override
+    public RoleAssignmentImpl define(String name) {
+        return wrapModel(name);
+    }
+
+    @Override
+    public GraphRbacManager manager() {
+        return this.manager;
+    }
+
+    @Override
+    public RoleAssignmentsInner inner() {
+        return manager().roleInner().roleAssignments();
+    }
+}