+++ sdk/sql/mgmt/src/main/java/com/azure/resourcemanager/sql/SloUsageMetricInterface.java
@@ -1,0 +1,21 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.resourcemanager.sql;
+
+import com.azure.core.annotation.Fluent;
+import com.azure.resourcemanager.resources.fluentcore.model.HasInner;
+import java.util.UUID;
+
+/** An immutable client-side representation of an Azure SQL database's SloUsageMetric. */
+@Fluent
+public interface SloUsageMetricInterface extends HasInner<SloUsageMetric> {
+    /** @return the serviceLevelObjective for SLO usage metric. */
+    ServiceObjectiveName serviceLevelObjective();
+
+    /** @return the serviceLevelObjectiveId for SLO usage metric. */
+    UUID serviceLevelObjectiveId();
+
+    /** @return inRangeTimeRatio for SLO usage metric. */
+    double inRangeTimeRatio();
+}