+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -1,10 +1,12 @@
 # Release History
 
 ## 1.0.0-beta.3 (Unreleased)
+- `USReceipt`, `USReceiptItem`, `USReceiptType` and `FormField{T}` types removed. Information about a `RecognizedReceipt` 
+must now be extracted from its `RecognizedForm`.
 - Rename parameters data and sourceUrl parameters found on methods for FormRecognizerClient to form and formUrl, respectively.
 - Rename parameters for receipt API methods to receipt and receiptUrl.
-- Raise `HttpResponseException` when a model with `ModelStatus.Invalid` is returned from the `beginTraining()` API's
-- Fix `HttpResponseException` to include the error object thrown on invalid analyze status for recognize API's
+- Raise `FormRecognizerException` when a model with `ModelStatus.Invalid` is returned from the `beginTraining()` API's
+- Raise `FormRecognizerException` when invalid analyze status returned from service for recognize API's
 - Update FormField property `transactionTime` on `USReceipt` to return `LocalTime` instead of `String`
 - Rename model `PageRange` to `FormPageRange`
 - Rename property `startPageNumber` to `firstPageNumber` and `endPageNumber` to `lastPageNumber` in model `PageRange`